+++ drivers/hetznercloud/destroy.go
@@ -9,37 +9,33 @@ import (
 	"strconv"
 
 	"github.com/drone/autoscaler"
+	"github.com/drone/autoscaler/logger"
 
 	"github.com/hetznercloud/hcloud-go/hcloud"
-	"github.com/rs/zerolog/log"
 )
 
 func (p *provider) Destroy(ctx context.Context, instance *autoscaler.Instance) error {
-	logger := log.Ctx(ctx).With().
-		Str("region", instance.Region).
-		Str("image", instance.Image).
-		Str("size", instance.Size).
-		Str("name", instance.Name).
-		Logger()
+	logger := logger.FromContext(ctx).
+		WithField("region", instance.Region).
+		WithField("image", instance.Image).
+		WithField("size", instance.Size).
+		WithField("name", instance.Name)
 
 	id, err := strconv.Atoi(instance.ID)
 	if err != nil {
 		return err
 	}
 
-	logger.Debug().
-		Msg("deleting instance")
+	logger.Debugln("deleting instance")
 
 	_, err = p.client.Server.Delete(ctx, &hcloud.Server{ID: id})
 	if err != nil {
-		logger.Error().
-			Err(err).
-			Msg("deleting instance failed")
+		logger.WithError(err).
+			Errorln("deleting instance failed")
 		return err
 	}
 
-	logger.Debug().
-		Msg("instance deleted")
+	logger.Debugln("instance deleted")
 
 	return nil
 }