+++ drivers/packet/create.go
@@ -10,8 +10,9 @@ import (
 	"time"
 
 	"github.com/drone/autoscaler"
+	"github.com/drone/autoscaler/logger"
+
 	"github.com/packethost/packngo"
-	"github.com/rs/zerolog/log"
 )
 
 func (p *provider) Create(ctx context.Context, opts autoscaler.InstanceCreateOpts) (*autoscaler.Instance, error) {
@@ -25,14 +26,13 @@ func (p *provider) Create(ctx context.Context, opts autoscaler.InstanceCreateOpt
 		return nil, err
 	}
 
-	logger := log.Ctx(ctx).With().
-		Str("project", p.project).
-		Str("facility", p.facility).
-		Str("billing", p.billing).
-		Str("plan", p.plan).
-		Str("os", p.os).
-		Str("hostname", p.hostname).
-		Logger()
+	logger := logger.FromContext(ctx).
+		WithField("project", p.project).
+		WithField("facility", p.facility).
+		WithField("billing", p.billing).
+		WithField("plan", p.plan).
+		WithField("os", p.os).
+		WithField("hostname", p.hostname)
 
 	cr := &packngo.DeviceCreateRequest{
 		HostName:     p.hostname,
@@ -44,14 +44,12 @@ func (p *provider) Create(ctx context.Context, opts autoscaler.InstanceCreateOpt
 		UserData:     buf.String(),
 	}
 
-	logger.Debug().
-		Msg("instance create")
+	logger.Debugln("instance create")
 
 	d, _, err := p.client.Devices.Create(cr)
 	if err != nil {
-		logger.Error().
-			Err(err).
-			Msg("cannot create instance")
+		logger.WithError(err).
+			Errorln("cannot create instance")
 		return nil, err
 	}
 
@@ -71,23 +69,20 @@ poller:
 	for {
 		select {
 		case <-ctx.Done():
-			logger.Debug().
-				Str("name", instance.Name).
-				Msg("cannot ascertain network")
+			logger.WithField("name", instance.Name).
+				Debugln("cannot ascertain network")
 
 			return instance, ctx.Err()
 		case <-time.After(interval):
 			interval = time.Minute
 
-			logger.Debug().
-				Str("name", instance.Name).
-				Msg("find instance network")
+			logger.WithField("name", instance.Name).
+				Debugln("find instance network")
 
 			d, _, err := p.client.Devices.Get(d.ID)
 			if err != nil {
-				logger.Error().
-					Err(err).
-					Msg("cannot find instance")
+				logger.WithError(err).
+					Errorln("cannot find instance")
 				return instance, err
 			}
 
@@ -105,10 +100,10 @@ poller:
 		}
 	}
 
-	logger.Debug().
-		Str("name", instance.Name).
-		Str("ip", instance.Address).
-		Msg("instance network ready")
+	logger.
+		WithField("name", instance.Name).
+		WithField("ip", instance.Address).
+		Debugln("instance network ready")
 
 	return instance, nil
 }