+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Use the new Docker runner image and deprecate the agent, by [@bradrydzewski](https://github.com/bradrydzewski).
+- Use logrus for logging instead of zerolog, by [@bradrydzewski](https://github.com/bradrydzewski).
 
 ## [1.4.3]
 ### Fixed